# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Payment Control Plane (PCP) API
  description: |
    Neutral payment orchestration platform. Provides one unified API for routing
    payments across multiple providers (Stripe, PayPal, Adyen, Razorpay).
    
    **This is NOT a payment gateway.** PCP never processes or stores raw card data.
  version: "0.7.0"
  contact:
    name: PCP Engineering
    email: engineering@pcp.dev
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.pcp.example.com
    description: Production
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...
paths:
...
  /api/v1/analytics/providers:
    get:
      tags: [Analytics]
      summary: Per-provider statistics
      operationId: getProviderStats
      parameters:
        - name: days
          in: query
          schema: { type: integer, default: 30 }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProviderStats" }
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProviderStats:
      type: object
      properties:
        provider_id: { type: string, format: uuid }
        provider_name: { type: string }
        total_charges: { type: integer }
        success_count: { type: integer }
        failure_count: { type: integer }
        total_amount: { type: integer, format: int64 }
        success_rate: { type: number, format: double }
        avg_latency_ms: { type: integer, format: int64 }